prop-settings-3: Refactor ColorInput to ColorGroup w/ new semantics (WIP)
include refactor _control array syntax, adding new _controlDef
