+++ app-templates/_default.template.toml
@@ -28,8 +28,8 @@ sourceColor = "#FFa500"
 _schemaVersion = "2.1"
 
 [_ui_defs]
-NodeType = { _control="in_color", label = "string", color = "string" }
-EdgeType = { _control="in_color", label = "string", color = "string" }
+NodeType = { label = "string", color = "string" }
+EdgeType = { label = "string", color = "string" }
 
 [_ui._groupMeta]
 label = "Global Settings"
@@ -92,7 +92,7 @@ exportLabel = { _control = "in_string", label = "Export Label", help = "Label us
 help = { _control = "in_string", label = "Help Text", help = "Help text shown to users" }
 hidden = { _control = "in_boolean", label = "Hide Field", help = "hide this info" }
 includeInGraphTooltip = { _control = "in_boolean", label = "Include In Graph Tooltip?", help = "display additional info" }
-options = { _control = "NodeType[]", label = "Node Types", help = "Define custom Node Types" }
+options = { _control = "in_colorgroup",_controlDef = "NodeType", label = "Node Types", help = "Define custom Node Types" }
 
 [_ui.nodeDefs.notes]
 _control = "composite"
@@ -191,7 +191,7 @@ displayLabel = { _control = "in_string", label = "label", help = "help|tooltip" 
 help = { _control = "in_string", label = "label", help = "help|tooltip" }
 isProvenance = { _control = "in_boolean", label = "label", help = "help|tooltip" }
 hidden = { _control = "in_boolean", label = "label", help = "help|tooltip" }
-options = { _control = "EdgeType[]", label = "Edge Types", help = "Define custom Edge Types" }
+options = { _control = "in_colorgroup", label = "Edge Types", help = "Define custom Edge Types" }
 
 [_ui.edgeDefs.notes]
 _control = "composite"
