+++ app/view/netcreate/components/MURColorInput.jsx
@@ -1,13 +1,12 @@
 /*///////////////////////////////// ABOUT \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\*\
 
   MUR Color Input Component
-  Handle a list of color options in a variable length array
+  Handle color item editing with color picker and label field
 
 \*\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\ * /////////////////////////////////////*/
 
 const React = require('react');
 const RSB = require('./react-settings-bridge');
-import ColorItemEdit from './MURColorItemEdit';
 const { ConsoleStyler } = require('ursys-min');
 
 /// CONSTANTS /////////////////////////////////////////////////////////////////
@@ -18,139 +17,222 @@ const PR = ConsoleStyler('InColor', 'TagBlue');
 
 /// STYLING OBJECTS ///////////////////////////////////////////////////////////
 /// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-const { inputStyle, opBtnStyle } = RSB.GetStyles();
-
-const arrayContainerStyle = {
-  border: '1px solid #ddd',
-  padding: '1rem',
-  backgroundColor: '#fff'
-};
+const { itemGrid, labelStyle, inputStyle, popupStyle, modColor, opBtnStyle } =
+  RSB.GetStyles();
 
 /// HELPER METHODS ////////////////////////////////////////////////////////////
 /// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-function m_ExtractColorArrayProps(controlData) {
+function m_ExtractColorProps(controlData) {
   if (typeof controlData !== 'object')
     return { error: `arg1 must be an object, got ${typeof controlData}` };
 
   const { groupName, propName, propField } = controlData;
-  const { sourceMeta, sourceData, itemDef } = controlData;
+  const { sourceMeta, sourceData } = controlData;
 
-  if (!Array.isArray(sourceData)) return { error: 'sourceData must be an array' };
+  const { label, tooltip, help } = sourceMeta;
+  const { labelKey, tooltipKey, helpKey, valueKey } = sourceMeta;
 
-  if (!itemDef || typeof itemDef !== 'object')
-    return { error: 'itemDef must be an object defining field types' };
+  // resolve useful ui data
+  let fValue, fLabel, fHelp, fTooltip, colorValue, labelValue;
+
+  // special case ... this is always a sourceData extraction
+  if (valueKey) fValue = sourceData[valueKey];
+  if (fValue === undefined) fValue = sourceData[propName] || sourceData;
+
+  // if labelKey exists, use it. Otherwise meta has to provide a label
+  if (labelKey) fLabel = sourceData[labelKey];
+  if (!fLabel) fLabel = label || '<label not in template>';
+
+  // if helpKey exists, use it. Otherwise meta has to provide a help text
+  if (helpKey) fHelp = sourceData[helpKey];
+  if (!fHelp) fHelp = help || '';
+
+  // if tooltipKey exists, use it. Otherwise meta has to provide a tooltip
+  if (tooltipKey) fTooltip = sourceData[tooltipKey];
+  if (!fTooltip) fTooltip = tooltip || fHelp || '';
+
+  // extract color and label from the item object
+  if (fValue) {
+    colorValue = fValue.color || '#000000';
+    labelValue = fValue.label || '';
+  }
 
   return {
     groupName,
     propName,
     propField,
-    sourceData,
-    itemDef,
-    sourceMeta
+    colorValue,
+    labelValue,
+    label: fLabel,
+    tooltip: fTooltip,
+    help: fHelp
   };
 }
 
-/// COLOR INPUT COMPONENT /////////////////////////////////////////////////////
+/// COLOR ITEM EDIT COMPONENT /////////////////////////////////////////////////
 /// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-/** Renders an array of color UI Objects */
+/** A ColorInput component */
 function ColorInput(props) {
+  // propName, groupName.propName, or groupName.propName.fieldName
   const { propDef } = props;
-  const { draft, hasLock } = React.useContext(RSB.SettingsContext);
-
+  const { draft, hasLock, dispatch } = React.useContext(RSB.SettingsContext);
+  // controlData contains what's needed to render this input component
   const controlData = RSB.GetDataForProp(draft.template, propDef);
-  const arrayProps = m_ExtractColorArrayProps(controlData);
+  const { colorValue, labelValue, label, tooltip, help } =
+    m_ExtractColorProps(controlData);
+  const defColorValue = '#000000';
+  const defLabelValue = '';
 
-  if (arrayProps.error) {
-    return (
-      <div style={{ color: 'red', padding: '1rem' }}>Error: {arrayProps.error}</div>
-    );
-  }
+  // declare reactive render state
+  // note that this only runs on the FIRST render, which is why
+  // we need the useEffect() to work around it
+  const [labelColor, setLabelColor] = React.useState('black');
+  const [tooltipStyle, setTooltipStyle] = React.useState({ ...popupStyle });
+  const templateColorValue = colorValue || defColorValue;
+  const templateLabelValue = labelValue || defLabelValue;
+  const [inputColorValue, setInputColorValue] = React.useState(
+    colorValue || defColorValue
+  );
+  const [inputLabelValue, setInputLabelValue] = React.useState(
+    labelValue || defLabelValue
+  );
 
-  const { groupName, propName, propField } = arrayProps;
-  const { sourceMeta, sourceData, itemDef } = arrayProps;
-  const isDisabled = !hasLock;
+  // this is a workaround for React's stupidity about dataflow, state
+  // retention with hooks, and other bullshit.
+  React.useEffect(() => {
+    if (!draft.pending) {
+      const { colorValue, labelValue } = m_ExtractColorProps(controlData);
+      setInputColorValue(colorValue || defColorValue);
+      setInputLabelValue(labelValue || defLabelValue);
+    }
+  }, [draft.pending]);
 
-  const { _control } = itemDef;
+  /// UI-SETTINGS INTEROP EVENT UPDATES ///
 
-  // Render color item editors for each item in the array
-  const ColorItemControls = sourceData.map((item, index) => {
-    const itemKey = `color-item-${index}`;
-    const itemPropDef = `${propDef}[${index}]`;
+  // send color data to draft object, which will trigger rerender
+  const submitColorToSettings = async newColor => {
+    const value = { color: newColor, label: inputLabelValue };
+    dispatch({
+      op: 'update',
+      propDef,
+      value
+    });
+  };
 
-    return (
-      <div
-        key={itemKey}
-        style={{ display: 'flex', alignItems: 'center', gap: '0.5rem' }}
-      >
-        <ColorItemEdit propDef={itemPropDef} />
-        <button
-          style={{
-            ...opBtnStyle,
-            backgroundColor: '#ff4444',
-            color: 'white',
-            fontSize: '0.8em',
-            padding: '0.2rem 0.5rem',
-            width: '60px'
-          }}
-          disabled={isDisabled}
-          onClick={() => {
-            // TODO: dispatch DELETE action
-            console.log(`Delete color item at ${itemPropDef}`);
-          }}
-        >
-          DELETE
-        </button>
-      </div>
-    );
-  });
+  // send label data to draft object, which will trigger rerender
+  const submitLabelToSettings = async newLabel => {
+    const value = { color: inputColorValue, label: newLabel };
+    dispatch({
+      op: 'update',
+      propDef,
+      value
+    });
+  };
 
-  return (
-    <div style={arrayContainerStyle}>
-      {ColorItemControls}
+  /// LOCAL EVENT UPDATES ///
+
+  // color changes will update the current value
+  const handleColorChange = event => {
+    const color = event.target.value;
+    setInputColorValue(color);
+    // TODO: dispatch color change
+    console.log(`Color change for ${propDef}:`, color);
+  };
+
+  // label changes will update the current value
+  const handleLabelTyping = event => {
+    const label = event.target.value;
+    setInputLabelValue(label);
+  };
+
+  // label key return will submit the value to draft object
+  const handleLabelEnterKey = async event => {
+    if (event.key === 'Enter') {
+      // TODO: dispatch label change
+      console.log(`Label change for ${propDef}:`, event.target.value);
+    }
+  };
+
+  // hovering over label will show tooltip
+  const handleTooltip = event => {
+    if (event.type === 'mouseover') {
+      const offset = RSB.EventTargetOffsetStyle(event);
+      const content = tooltip || '';
+      setTooltipStyle({
+        ...popupStyle,
+        ...offset,
+        display: 'block',
+        content
+      });
+      setLabelColor('maroon');
+    } else if (event.type === 'mouseout') {
+      setTooltipStyle({ ...popupStyle });
+      setLabelColor('black');
+    }
+  };
+
+  /// SUB RENDER ///
+
+  // conditional flags based on template values
+  const colorMod = templateColorValue !== inputColorValue;
+  const labelMod = templateLabelValue !== inputLabelValue;
+  const mod = colorMod || labelMod;
+  const bgColor = mod ? modColor : 'white';
+
+  // render either enabled or disabled based on hasLock
+  const InputField = hasLock ? (
+    <div style={{ display: 'flex', alignItems: 'center', gap: '0.5rem' }}>
+      <input
+        type="color"
+        value={inputColorValue}
+        style={{
+          width: '40px',
+          height: '30px',
+          border: '1px solid #ccc',
+          borderRadius: '3px',
+          backgroundColor: bgColor
+        }}
+        onChange={handleColorChange}
+      />
+      <input
+        type="text"
+        value={inputLabelValue}
+        placeholder="Label"
+        style={{
+          ...inputStyle,
+          minWidth: '120px',
+          backgroundColor: bgColor
+        }}
+        onKeyDown={handleLabelEnterKey}
+        onBlur={e => {
+          // TODO: dispatch label change on blur
+          console.log(`Label blur for ${propDef}:`, e.target.value);
+        }}
+        onChange={handleLabelTyping}
+      />
+    </div>
+  ) : (
+    <div style={{ display: 'flex', alignItems: 'center', gap: '0.5rem' }}>
       <div
         style={{
-          display: 'flex',
-          justifyContent: 'space-between',
-          alignItems: 'center',
-          marginTop: '1rem',
-          paddingTop: '1rem',
-          borderTop: '1px solid #ddd'
+          width: '20px',
+          height: '20px',
+          backgroundColor: inputColorValue,
+          border: '1px solid #ccc',
+          borderRadius: '3px'
         }}
-      >
-        <div style={{ display: 'flex', alignItems: 'center', gap: '0.5rem' }}>
-          <label>SORT:</label>
-          <select
-            style={inputStyle}
-            disabled={isDisabled}
-            defaultValue="a-z"
-            onChange={e => {
-              // TODO: dispatch SORT action
-              console.log(`Sort ${propDef} by:`, e.target.value);
-            }}
-          >
-            <option value="a-z">A-Z</option>
-            <option value="z-a">Z-A</option>
-          </select>
-        </div>
-        <button
-          style={{
-            ...opBtnStyle,
-            backgroundColor: '#4CAF50',
-            color: 'white',
-            fontSize: '0.8em',
-            padding: '0.2rem 0.5rem',
-            width: '60px',
-            marginRight: '1rem'
-          }}
-          disabled={isDisabled}
-          onClick={() => {
-            // TODO: dispatch ADD action
-            console.log(`Add new color item to ${propDef}`);
-          }}
-        >
-          ADD
-        </button>
-      </div>
+      />
+      <span>{inputLabelValue || 'Unlabeled'}</span>
+      <span style={{ color: 'gray', fontSize: '0.8em' }}>({inputColorValue})</span>
+    </div>
+  );
+
+  /// RENDER ///
+
+  return (
+    <div style={itemGrid}>
+      {InputField}
+      {tooltip && <div style={tooltipStyle}>{tooltip}</div>}
     </div>
   );
 }
